removed observer pattern from dryer listener
The previously implemented observer pattern was removed from the DryerListener class. This pattern handled the output of the listener. The output will be now directly diverted to the parent class of the listener. Moreover, an activity can now communicate with its function buttons fragment to open the door and hide the relevant button.